% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.zinb.R
\name{apply.zinb}
\alias{apply.zinb}
\title{Apply ZINB on dataset.
This is a general function which applies ZINB on a given dataset.}
\usage{
apply.zinb(dataset, group = NULL)
}
\arguments{
\item{dataset}{Numeric count matrix of read counts.}

\item{group}{Vector or factor giving the experimental group/condition for each sample/library in the pilot dataset. If not provided, groups are assumed equally divided.}

\item{filename}{Filename of the dataset; ZINB results on that dataset will be named accordingly}
}
\value{
Applies ZINB to given dataset, and returns results table
}
\description{
Apply ZINB on dataset.
This is a general function which applies ZINB on a given dataset.
}
\examples{

}

