% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downSample.R
\name{downSample}
\alias{downSample}
\title{Down-sample RNA-seq data.
Down-sample full dataset to a smaller dataset of specified size.}
\usage{
downSample(dataset, group = NULL, nreps, depth, nsims = 20)
}
\arguments{
\item{dataset}{Numeric count matrix; full dataset from which to down-sample. Note that for my purposes, this full dataset is simulated.}

\item{group}{Vector or factor giving the experimental group/condition for each sample/library in the full dataset.}

\item{nreps}{Desired number of replicates per group.}

\item{depth}{Desired depth (library size) in millions for each sample.}

\item{nsims}{Number of simulations to run, i.e. number of datasets to generate of this size.}
}
\value{
Doesn't return anything, just saves the down-sampled datasets as RData.
}
\description{
Down-sample RNA-seq data.
Down-sample full dataset to a smaller dataset of specified size.
}
\examples{

}

