% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout.fun.R
\name{dropout.fun}
\alias{dropout.fun}
\title{Impose dropouts. 
Impose dropouts}
\usage{
dropout.fun(object, pDropout, drop.method, drop.lambda, seed)
}
\arguments{
\item{object}{DGEList object.}

\item{pDropout}{Numeric vector of dropout probabilities to consider.}

\item{drop.method}{Dropout method, with two options: zero-inflation, or low-rate Poisson.}

\item{drop.lambda}{Mean of low-rate Poisson for the Poisson-dropout scenario.}

\item{seed}{Optional seed, if it desired to replicate the simulation.}
}
\value{
Saves attributes of dropout imposition on simulated data to DGEList object: dropout-imposed count matrix, gene-wise probability of dropout, logical matrix indicating dropout observations.
}
\description{
Impose dropouts. 
Impose dropouts
}
\examples{

}
\keyword{internal}

