% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataset.R
\name{getDataset}
\alias{getDataset}
\title{Extract real dataset parameters. 
Extract Negative Binomial parameters from real dataset.}
\usage{
getDataset(counts, drop.extreme.dispersion = 0.1, drop.low.lambda = TRUE)
}
\arguments{
\item{counts}{Numeric matrix of read counts from which to extract parameters.}

\item{drop.extreme.dispersion}{Proportion of extreme dispersions to drop.}

\item{drop.low.lambda}{Logical, whether to drop low lambdas.}
}
\value{
List of parameters sampled from the real dataset: vector of gene-wise average log-CPM, vector of gene-wise dispersions, library sizes, number of genes.
}
\description{
Extract real dataset parameters. 
Extract Negative Binomial parameters from real dataset.
}
\examples{

}
\keyword{internal}

