% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.R
\name{makePlots}
\alias{makePlots}
\title{Custom multiplot function for ggplot2}
\usage{
makePlots(reps, depths, plotROC = TRUE, plotDE = TRUE, plotPower = TRUE,
  cols = 2, width = 10, height = 10)
}
\arguments{
\item{reps}{Numerical vector of replicates per gorup to plot}

\item{depths}{Numerical vector of depths to plot, in millions}

\item{plotROC}{Logical (default TRUE), whether to make ROC plots}

\item{plotDE}{Logical (default TRUE), whether to make numDE plots}

\item{plotPower}{Logical (default TRUE), whether to make power plots}

\item{cols}{Number of columns to plot (default)}

\item{width, }{height The width and height of the graphics region in inches; default values are 10}
}
\value{
Makes plots of ROC curves, number of detected DE features, and power, all saved as PDFs
}
\description{
Custom multiplot function for ggplot2
}
\examples{

}

