% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Custom multiplot function for ggplot2
#param ... ggplot objects}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{List of ggplot objects}

\item{cols}{Number of columns}

\item{layout}{If NULL, then use 'cols' to determine layout}
}
\value{
Renders multiple ggplot plots in one image
}
\description{
Custom multiplot function for ggplot2
#param ... ggplot objects
}
\examples{

}

