% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordDE.R
\name{recordDE}
\alias{recordDE}
\title{Record the number of differentially expressed genes found by each of the edgeR results}
\usage{
recordDE(reps, depths, nsims)
}
\arguments{
\item{reps}{Numerical vector of replicates per group considered.}

\item{depths}{Numerical vector of depths considered (e.g. c(0.5, 1, 1.5)*1e6)}

\item{nsims}{Number of simulations that were run per reps and depths}
}
\value{
Doesn't return anything, just writes results to file as "recordDE.txt". Table contains the number of DE genes for each reps x depth x sim dataset on which edgeR was applied.
}
\description{
Record the number of differentially expressed genes found by each of the edgeR results
}
\examples{

}

