% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordOverlap.R
\name{recordOverlap}
\alias{recordOverlap}
\title{Records overlaps with gold standard DE gene list, and gets accuracy measures}
\usage{
recordOverlap(reps, depths, nsims)
}
\arguments{
\item{reps}{Numerical vector of replicates per group considered.}

\item{depths}{Numerical vector of depths considered (e.g. c(0.5, 1, 1.5)*1e6)}

\item{nsimss}{Number of simulations that were run per reps and depths}
}
\description{
Records overlaps with gold standard DE gene list, and gets accuracy measures
}
\examples{

}

