% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordPlotData.R
\name{recordPlotData}
\alias{recordPlotData}
\title{Record the number of differentially expressed genes found by each of the edgeR results}
\usage{
recordPlotData(reps, depths, nsims)
}
\arguments{
\item{reps}{Numerical vector of replicates per group considered.}

\item{depths}{Numerical vector of depths considered, in millions}

\item{nsims}{Number of simulations that were run per reps and depths}
}
\value{
Writes plotting data to file
}
\description{
Record the number of differentially expressed genes found by each of the edgeR results
}
\examples{

}

