% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.fun.R
\name{sample.fun}
\alias{sample.fun}
\title{Re-sample real dataset parameters. 
Re-sample parameters extracted from real dataset, to be used for the simulation dataset.}
\usage{
sample.fun(object, seed)
}
\arguments{
\item{object}{DGEList object.}

\item{seed}{Optional seed, if it desired to replicate the simulation.}
}
\value{
Saves re-sampled gene-wise lambdas and dispersions to DGEList object.
}
\description{
Re-sample real dataset parameters. 
Re-sample parameters extracted from real dataset, to be used for the simulation dataset.
}
\examples{

}
\keyword{internal}

