% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fun.R
\name{sim.fun}
\alias{sim.fun}
\title{Simulate scRNA-seq dataset.
Uses previously re-sampled parameters to simulate data according to Negative Binomial distribution.}
\usage{
sim.fun(object, seed)
}
\arguments{
\item{object}{DGEList object.}

\item{seed}{Optional seed, if it desired to replicate the simulation.}
}
\value{
Saves simulated DE data to DGEList object.
}
\description{
Simulate scRNA-seq dataset.
Uses previously re-sampled parameters to simulate data according to Negative Binomial distribution.
}
\examples{

}
\keyword{internal}

