% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.scRNAseq.R
\name{sim.scRNAseq}
\alias{sim.scRNAseq}
\title{Simulating scRNA-seq expression data}
\usage{
sim.scRNAseq(dataset, nlibs = NULL, nTags = NULL, groups = NULL,
  beta = NULL, drop.low.lambda = TRUE, drop.extreme.dispersion = 0.1,
  lib.size = NULL, flibs = c(0.7, 1.3), verbose = TRUE, seed = NULL)
}
\arguments{
\item{dataset}{Numeric matrix of read counts from which to extract parameters.}

\item{nlibs}{Desired number of replicates per group.}

\item{nTags}{Desired number of genes. If not provided, simulated dataset will contain as many genes as the real dataset.}

\item{groups}{Vector indicating membership in at most 2 groups}

\item{beta}{Set of coefficients for the model matrix (must have same number of columns as mod)}

\item{drop.low.lambda}{Logical, whether to drop low lambdas.}

\item{drop.extreme.dispersion}{Proportion of extreme dispersions to drop.}

\item{lib.size}{Numeric vector giving the total count (sequence depth) for each library. If not provided, library sizes are extracted from the real dataset.}

\item{flibs}{Some wiggle room for the library sizes extracted from real
dataset.}

\item{verbose}{Logical, whether to print simulation progress to screen.}

\item{seed}{Optional seed, if it desired to replicate the simulation.}

\item{add.dropout}{Logical, whether to impose dropouts.}

\item{pDropout}{Numeric vector of dropout probabilities to consider. Will only be used if \code{add.dropout=TRUE}.}

\item{drop.method}{Dropout method, with two options: zero-inflation (default), or low-rate Poisson. Will only be used if \code{add.dropout=TRUE}.}

\item{drop.lambda}{Mean of low-rate Poisson for the Poisson-dropout scenario. Will only be used if \code{add.dropout=TRUE} and \code{drop.method="poisson"}.}
}
\value{
DGElist object containing all simulation settings and results, including the count matrix and dropout-imposed count matrix.
}
\description{
Extract parameters from real dataset to simulate scRNA-seq expression data with differentially expressed genes; options for imposing dropouts on the data.
}
\examples{
# Simulate toy dataset with 3 reps per group and 20 genes
# data(prostatedata)
#simData <- sim.scRNAseq(dataset=prostatedata, nreps=3, nTags=20, drop.method=c("zero"))
NULL
}

