% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.scRNAseq.R
\name{sim.scRNAseq}
\alias{sim.scRNAseq}
\title{Simulating scRNA-seq expression data 
Extract parameters from real dataset to simulate scRNA-seq expression data with differentially expressed genes; options for imposing dropouts on the data.}
\usage{
sim.scRNAseq(dataset, nreps, group = NULL, nTags = NULL, lib.size = NULL,
  flibs = c(0.7, 1.3), pDiff = 0.1, pUp = 0.5, logFC = c(0.25, 1.5),
  drop.low.lambda = TRUE, drop.extreme.dispersion = 0.1,
  add.dropout = TRUE, pDropout = c(0.1, 0.2, 0.5, 0.8),
  drop.method = c("zero", "poisson"), drop.lambda = 1, verbose = TRUE,
  seed = NULL)
}
\arguments{
\item{dataset}{Numeric matrix of read counts from which to extract
parameters.}

\item{nreps}{Desired number of replicates per group.}

\item{group}{Vector or factor giving the experimental group/condition for
each sample/library.}

\item{nTags}{Desired number of genes. If not provided, simulated dataset will
contain as many genes as the real dataset.}

\item{lib.size}{Numeric vector giving the total count (sequence depth) for
each library. If not provided, library sizes are extracted from the real
dataset.}

\item{flibs}{Some wiggle room for the library sizes extracted from real
dataset.}

\item{pDiff}{Proportion of differentially expressed genes.}

\item{pUp}{Proportion of differentially expressed genes that are upregulated.}

\item{logFC}{Desired log-fold changes for differentially expressed genes. If single number, all DE genes will exhibit constant logFC. If vector of two numbers provided (default \code{c(0.25, 1.5)}), logFC of DE genes will be drawn uniformly within that range.}

\item{drop.low.lambda}{Logical, whether to drop low lambdas.}

\item{drop.extreme.dispersion}{Proportion of extreme dispersions to drop.}

\item{add.dropout}{Logical, whether to impose dropouts.}

\item{pDropout}{Numeric vector of dropout probabilities to consider. Will only be used if \code{add.dropout=TRUE}.}

\item{drop.method}{Dropout method, with two options: zero-inflation, or low-rate Poisson. Will only be used if \code{add.dropout=TRUE}.}

\item{drop.lambda}{Mean of low-rate Poisson for the Poisson-dropout scenario. Will only be used if \code{add.dropout=TRUE} and \code{drop.method="poisson"}.}

\item{verbose}{Logical, whether to print simulation progress to screen.}

\item{seed}{Optional seed, if it desired to replicate the simulation.}
}
\value{
A numeric count matrix of desired size, simulated with parameters extracted from a real dataset, containing differentially expressed genes and (optionally imposed) dropouts.
}
\description{
Simulating scRNA-seq expression data 
Extract parameters from real dataset to simulate scRNA-seq expression data with differentially expressed genes; options for imposing dropouts on the data.
}
\examples{
# Simulate toy dataset with 3 reps per group and 20 genes
# data(prostatedata)
#simData <- sim.scRNAseq(dataset=prostatedata, nreps=3, nTags=20, drop.method=c("zero"))
NULL
}

